server/projects/main/requirements.txt
# Main Server Requirements
# for server deploy
gunicorn==20.1.0

# for django
Django==3.1.12

# for django using mysql server
pymysql==1.0.2

# for django template
django-widget-tweaks==1.4.8
Jinja2==2.11.3

# for celery
celery==4.3.0
vine==1.3.0
django-celery-results==1.1.2
redis==3.3.11
celery-redbeat==0.13.0
flower==0.9.3

# for rest api
djangorestframework==3.11.1
coreapi==2.3.3
Markdown==3.3.4
Pygments==2.8.1
django-filter==2.4.0
django-cors-headers==3.7.0

# for Document
drf-yasg==1.17.1

# for password
pyaes==1.6.1
pycryptodome==3.9.8
cryptography==2.9.2
rsa==4.7.2

# for django permission
django-guardian==2.3.0

# for file upload and download
urllib3==1.26.7

# for database import and export module
django-import-export==2.5.0

# for exception log
sentry-sdk==0.10.2

# for monitor
django-prometheus==2.0.0

# for login
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
